requirements.txt
aiogram==3.0.0
aiohttp==3.8.6
openai==0.27.2
python-dotenv==1.0.0
async-timeout==4.0.3
beautifulsoup4==4.11.2
psycopg2==2.9.6
python-docx==0.8.11
pdfplumber==0.9.0

